Refactor Rejigs builder: split character and build methods
Moved character-related and build-related methods from Rejigs.cs into new partial classes: Characters.cs and Building.cs. Updated tests to match new class names and added/expanded test cases for character methods. This improves code organization and test coverage.
