+++ Rejigs.Tests/CharactersTests.cs
@@ -8,6 +8,7 @@ public class Character
         public void AnyDigit_MatchesSingleDigit()
         {
             var regex = Rejigs.Create().AnyDigit().Build();
+            
             Assert.That("5", Does.Match(regex));
             Assert.That("a", Does.Not.Match(regex));
         }
@@ -16,7 +17,10 @@ public class Character
         public void AnyNonDigit_MatchesSingleNonDigit()
         {
             var regex = Rejigs.Create().AnyNonDigit().Build();
+            
             Assert.That("a", Does.Match(regex));
+            Assert.That("_", Does.Match(regex));
+            Assert.That("$", Does.Match(regex));
             Assert.That("5", Does.Not.Match(regex));
         }
 
@@ -24,8 +28,10 @@ public class Character
         public void AnyLetterOrDigit_MatchesWordCharacter()
         {
             var regex = Rejigs.Create().AnyLetterOrDigit().Build();
+            
             Assert.That("a", Does.Match(regex));
             Assert.That("_", Does.Match(regex));
+            Assert.That("1", Does.Match(regex));
             Assert.That("$", Does.Not.Match(regex));
         }
 
@@ -35,12 +41,14 @@ public class Character
             var regex = Rejigs.Create().AnyNonLetterOrDigit().Build();
             Assert.That("$", Does.Match(regex));
             Assert.That("a", Does.Not.Match(regex));
+            Assert.That("1", Does.Not.Match(regex));
         }
 
         [Test]
         public void AnySpace_MatchesWhitespace()
         {
             var regex = Rejigs.Create().AnySpace().Build();
+            
             Assert.That(" ", Does.Match(regex));
             Assert.That("\t", Does.Match(regex));
             Assert.That("a", Does.Not.Match(regex));
@@ -50,14 +58,17 @@ public class Character
         public void AnyNonSpace_MatchesNonWhitespace()
         {
             var regex = Rejigs.Create().AnyNonSpace().Build();
+            
             Assert.That("a", Does.Match(regex));
             Assert.That(" ", Does.Not.Match(regex));
+            Assert.That("\t", Does.Not.Match(regex));
         }
 
         [Test]
         public void AnyCharacter_MatchesAnyCharacter()
         {
             var regex = Rejigs.Create().AnyCharacter().Build();
+            
             Assert.That("a", Does.Match(regex));
             Assert.That("1", Does.Match(regex));
             Assert.That("$", Does.Match(regex));
@@ -68,6 +79,7 @@ public class Character
         public void AnyOf_MatchesCharactersInSet()
         {
             var regex = Rejigs.Create().AnyOf("abc").Build();
+            
             Assert.That("a", Does.Match(regex));
             Assert.That("b", Does.Match(regex));
             Assert.That("c", Does.Match(regex));
@@ -78,9 +90,11 @@ public class Character
         public void AnyInRange_MatchesCharactersInRange()
         {
             var regex = Rejigs.Create().AnyInRange('a', 'f').Build();
+            
             Assert.That("a", Does.Match(regex));
             Assert.That("d", Does.Match(regex));
             Assert.That("f", Does.Match(regex));
+            Assert.That("af", Does.Match(regex));
             Assert.That("g", Does.Not.Match(regex));
         }
 
@@ -88,7 +102,9 @@ public class Character
         public void AnyExcept_MatchesCharactersNotInSet()
         {
             var regex = Rejigs.Create().AnyExcept("xyz").Build();
+            
             Assert.That("a", Does.Match(regex));
+            Assert.That("aq", Does.Match(regex));
             Assert.That("x", Does.Not.Match(regex));
             Assert.That("y", Does.Not.Match(regex));
         }
